Smartphone class split into definition and implementation
