+++ smartphone.cpp
@@ -1,23 +1,11 @@
 #include <iostream>
 #include <string>
+#include <smartphone.h>
 
 class Screen
 {
 public:
   int glass;
-};
-
-
-class SmartPhone
-{
-public:
-  void dialNumber(std::string number);
-  void call();
-  void endCall();
-  void heySiri(std::string wakeUpNotice);
-  
-private:
-  Screen screen;
 };
 
 
@@ -39,16 +27,5 @@ void SmartPhone::endCall()
 void SmartPhone::heySiri(std::string wakeUpNotice)
 {
   std::cout << "How can I help you?" << std::endl;
-}
-
-int main()
-{
-  SmartPhone nokia;
-  nokia.dialNumber("0612345678");
-  nokia.call();
-  nokia.endCall();
-  nokia.heySiri("Hello Siri");
-
-  return 0;
 }
